refactor: Rename DTOs UserSeller folder to User
