+++ desafio_spring/src/main/java/com/desafiospringboot/DTOs/User/UserDTO.java
@@ -1,10 +1,9 @@
-package com.desafiospringboot.DTOs.UserSeller;
+package com.desafiospringboot.DTOs.User;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 import com.desafiospringboot.Entities.User;
-import com.desafiospringboot.Entities.UserSeller;
 
 public class UserDTO {
 
@@ -12,7 +11,6 @@ public class UserDTO {
 	private String userName;
 
 	public UserDTO() {
-
 	}
 
 	public UserDTO(User user ) {
