+++ desafio_spring/src/main/java/com/desafiospringboot/DTOs/User/UserFollowedPostsDTO.java
@@ -1,4 +1,4 @@
-package com.desafiospringboot.DTOs;
+package com.desafiospringboot.DTOs.User;
 
 import com.desafiospringboot.DTOs.Post.PostDTO;
