+++ desafio_spring/src/main/java/com/desafiospringboot/DTOs/User/UserSellerFollowedListDTO.java
@@ -1,24 +1,20 @@
-package com.desafiospringboot.DTOs.UserSeller;
+package com.desafiospringboot.DTOs.User;
 
 
 import com.desafiospringboot.Entities.UserClient;
 import com.desafiospringboot.Entities.UserSeller;
 
 import javax.validation.constraints.NotNull;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
-public class UserSellerFollowedDTO {
+public class UserSellerFollowedListDTO {
     @NotNull
     private int userId;
     @NotNull
     private String userName;
     private List<UserDTO> followed;
 
-    
-
-    public UserSellerFollowedDTO(int userId, String userName, List<UserDTO> followed) {
+    public UserSellerFollowedListDTO(int userId, String userName, List<UserDTO> followed) {
         this.userId = userId;
         this.userName = userName;
         this.followed = followed;
@@ -28,7 +24,7 @@ public class UserSellerFollowedDTO {
         return userId;
     }
 
-    public String getuserName() {
+    public String getUserName() {
         return userName;
     }
 
@@ -36,7 +32,7 @@ public class UserSellerFollowedDTO {
         return followed;
     }
 
-    public static UserSellerFollowedDTO convert (List<UserSeller> listSellers, UserClient user) { 	
-        return new UserSellerFollowedDTO(user.getId(), user.getName(), UserDTO.convert(listSellers));
+    public static UserSellerFollowedListDTO convert (List<UserSeller> listSellers, UserClient user) {
+        return new UserSellerFollowedListDTO(user.getId(), user.getName(), UserDTO.convert(listSellers));
     }
 }
