+++ desafio_spring/src/main/java/com/desafiospringboot/DTOs/User/UserSellerPromoPostDTO.java
@@ -1,4 +1,4 @@
-package com.desafiospringboot.DTOs.UserSeller;
+package com.desafiospringboot.DTOs.User;
 
 import com.desafiospringboot.DTOs.Post.PostPromoDTO;
 import com.desafiospringboot.Entities.Post;
